requirements-production.txt
absl-py==2.3.1
accelerate==1.6.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.10.0
attrs==25.3.0
audioread==3.0.1
av==14.4.0
babel==2.17.0
bangla==0.0.5
beautifulsoup4==4.13.5
blinker==1.9.0
blis==1.2.1
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
Brotli==1.1.0
catalogue==2.0.10
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.22.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.2
coqpit==0.0.17
cryptography==45.0.7
cycler==0.12.1
cymem==2.0.11
Cython==3.1.3
dateparser==1.1.8
decorator==5.2.1
diffusers==0.35.1
dnspython==2.8.0
docopt==0.6.2
EbookLib==0.19
einops==0.8.1
encodec==0.1.1
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.19.1
Flask==3.1.2
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.9.0
g2pkk==0.1.2
google-crc32c==1.7.1
gradio==5.44.1
gradio_client==1.12.1
groovy==0.1.2
grpcio==1.74.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
hangul-romanize==0.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.7.0
inflect==7.5.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.2
jsonlines==1.2.0
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
lxml==6.0.1
marisa-trie==1.3.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
ml_collections==1.1.0
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
murmurhash==1.0.13
networkx==2.8.8
nltk==3.9.1
num2words==0.5.14
numba==0.61.2
numpy==1.26.4
orjson==3.11.3
packaging==25.0
pandas==2.2.2
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
preshed==3.0.10
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.2
pylibsrtp==0.12.0
pynndescent==0.5.13
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypdf==6.0.0
pypinyin==0.55.0
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
ruff==0.12.12
safehttpx==0.1.6
safetensors==0.6.2
scikit-learn==1.4.2
scipy==1.14.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.47.3
SudachiDict-core==20250825
SudachiPy==0.6.10
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
thinc==8.3.4
threadpoolctl==3.6.0
tokenizers==0.21.4
tomlkit==0.13.3
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
trainer==0.0.36
transformers==4.51.3
TTS==0.22.0
typeguard==4.4.4
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
umap-learn==0.5.9.post2
Unidecode==1.4.0
urllib3==2.5.0
uvicorn==0.35.0
-e git+https://github.com/akadoubleone/VibeVoice-Community@f412a57368452f7eb7d997ddfe39f2dd60e6bb09#egg=vibevoice
wasabi==1.1.3
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
